pyproject.toml
[project]
name = "mcp-brandfetch"
version = "0.2.0"
description = "MCP server for Brandfetch brand data API"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "DJ", email = "dj@example.com" }
]
license = { text = "MIT" }
keywords = ["mcp", "brandfetch", "brand-data", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "cachetools>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
redis = [
    "redis>=4.0.0",
]

[project.scripts]
mcp-brandfetch = "brandfetch_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/brandfetch_mcp"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration (deselect with '-m \"not integration\"')",
]
testpaths = ["tests"]
